+++ WEB/Src/Web/Web.Tests/ClientIpHeaderTelemetryModuleTest.cs
@@ -80,22 +80,25 @@
         {
             var module = new TestableClientIpHeaderTelemetryInitializer();
             var requestTelemetry = module.PlatformContext.CreateRequestTelemetryPrivate();
-            
+
             module.Initialize(new EventTelemetry());
 
             Assert.AreEqual("127.0.0.1", requestTelemetry.Context.Location.Ip);
         }
 
         [TestMethod]
-        public void InitializeSetsLocationIpToUserHostAddressFromXForwarderHeader()
+        [DataRow("1.2.3.4", "1.2.3.4")]
+        [DataRow("[::1]:80", "::1")]
+        [DataRow("0:0:0:0:0:0:0:1", "::1")]
+        public void InitializeSetsLocationIpToUserHostAddressFromXForwarderHeader(string value, string expected)
         {
-            var dictionary = new Dictionary<string, string> { { "X-Forwarded-For", "1.2.3.4" } };
+            var dictionary = new Dictionary<string, string> { { "X-Forwarded-For", value } };
             var module = new TestableClientIpHeaderTelemetryInitializer(dictionary);
             var telemetry = new TraceTelemetry();
 
             module.Initialize(telemetry);
 
-            Assert.AreEqual("1.2.3.4", telemetry.Context.Location.Ip);
+            Assert.AreEqual(expected, telemetry.Context.Location.Ip);
         }
 
         [TestMethod]
@@ -104,7 +107,7 @@
             var dictionary = new Dictionary<string, string> { { "X-Forwarded-For", "1.2.3.4" } };
             var module = new TestableClientIpHeaderTelemetryInitializer(dictionary);
             var requestTelemetry = module.PlatformContext.CreateRequestTelemetryPrivate();
-            
+
             module.Initialize(new TraceTelemetry("trace"));
 
             Assert.AreEqual("1.2.3.4", requestTelemetry.Context.Location.Ip);